Add textures to your Crowd with a GUI
Use this code to create a new tool in your shelf. Once your Crowd sim is done and cached, select the Geometry node where the cache is stored and launch the tool. A GUI will open where you can choose a texture file for each Agent Shape. It will then generate a «Material Stylesheet» and apply all the materials for you.