Browse through your characters folder and import them as Agents
Use this code to create a new tool in your shelf. When you launch the tool, it will look for .FBX files in your Agent Directory and list them in a UI. You can then pick one or more characters from that list, and the following nodes will be generated for each of them: Agent, Agent Clip, Agent Layer and Agent Prep.